Add venn_segment_packed
This old, uncommitted modification probably does not belong in
PackedArray but I was thrilled to discover these magic 11 lines of code
I thought were lost. So, I'm going to commit them and may extract them
later. They more properly belong in a class that uses this package, not
in this package itself.